core_app/requirements.txt
pandas
pytube
opencv-python
numpy
pytesseract
python-dotenv
psutil
tqdm
faster-whisper
youtube-transcript-api
langchain
langchain-openai
spacy
firebase-admin